feat: implement increment goal functionality with API integration and background processing

+++ server/internal/model/goals_model.go
@@ -9,6 +9,6 @@ type Goals struct {
 	TargetDays    int                `bson:"targetDays" json:"targetDays"`
 	Category      string             `bson:"category" json:"category"`
 	Done          bool               `bson:"done" json:"done"`
-	CurrentTarget int                `bson:"currentTarget" json:"currentTarget"`
+	CurrentTarget int64                `bson:"currentTarget" json:"currentTarget"`
 	WorkspaceId   primitive.ObjectID `bson:"workspaceId" json:"workspaceId"`
 }
